requirements.txt
aiohttp[speedups] >= 3.5.4
asyncurban >= 0.3.3
beautifulsoup4 >= 4.8.1
covid >= 2.2.9
cowpy >= 1.1.0
dill >= 0.3.1.1
gitpython >= 3.0.3
hachoir >= 3.0a5
heroku3 >= 3.4.0
hiredis >= 1.0.0
meval >= 2.5
pillow >= 8.1.0
pyaesni >= 0.16
redis >= 3.3.10
requests >= 2.22.0
speedtest-cli >= 2.1.2
telethon >= 1.21.1
wikipedia >= 1.4.0
yt-dlp >= 2021.11.10.1
